Split delete into own slices
