+++ ui/dashboards/src/context/DashboardProvider/panel-group-editor-slice.ts
@@ -11,11 +11,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { getPanelKeyFromRef } from '@perses-dev/core';
 import { StateCreator } from 'zustand';
 import { Middleware } from './common';
-import { PanelGroupSlice, PanelGroupDefinition, PanelGroupId, mapPanelToPanelGroups } from './panel-group-slice';
-import { PanelSlice } from './panel-slice';
+import { PanelGroupSlice, PanelGroupDefinition, PanelGroupId } from './panel-group-slice';
 
 export interface PanelGroupEditor {
   mode: 'Add' | 'Edit';
@@ -29,13 +27,8 @@ export interface PanelGroupEditorValues {
   isCollapsed: boolean;
 }
 
-export interface DeletePanelGroupDialog {
-  panelGroupId: PanelGroupId;
-  panelGroupName?: string;
-}
-
 /**
- * Slice that handles the visual editor state and related actions for Panel Groups (i.e. add, edit, delete).
+ * Slice that handles the visual editor state and related actions for adding or editing Panel Groups.
  */
 export interface PanelGroupEditorSlice {
   /**
@@ -52,21 +45,11 @@ export interface PanelGroupEditorSlice {
    * Opens the panel group editor to edit an existing panel group.
    */
   openEditPanelGroup: (panelGroupId: PanelGroupId) => void;
-
-  deletePanelGroupDialog?: DeletePanelGroupDialog;
-
-  /**
-   * Delete panel group and all the panels within the group
-   */
-  deletePanelGroup: (panelGroupId: PanelGroupId) => void;
-
-  openDeletePanelGroupDialog: (panelGroupId: PanelGroupId) => void;
-  closeDeletePanelGroupDialog: () => void;
 }
 
 export const createPanelGroupEditorSlice: StateCreator<
-  // Actions in here need to modify both Panels and Panel Groups state
-  PanelGroupEditorSlice & PanelGroupSlice & PanelSlice,
+  // Actions in here need to modify Panel Group state
+  PanelGroupEditorSlice & PanelGroupSlice,
   Middleware,
   [],
   PanelGroupEditorSlice
@@ -140,49 +123,4 @@ export const createPanelGroupEditorSlice: StateCreator<
       draft.panelGroupEditor = editor;
     });
   },
-
-  deletePanelGroup(panelGroupId) {
-    const { panelGroups, panelGroupIdOrder } = get();
-    const group = panelGroups[panelGroupId];
-    const idIndex = panelGroupIdOrder.findIndex((id) => id === panelGroupId);
-    if (group === undefined || idIndex === -1) {
-      throw new Error(`Panel group ${panelGroupId} not found`);
-    }
-
-    // Get the panel keys for all the panel items in the group we're going to delete
-    const panelKeys = group.items.map((item) => getPanelKeyFromRef(item.content));
-
-    set((draft) => {
-      // Delete the panel group which also deletes all its items
-      delete draft.panelGroups[panelGroupId];
-      draft.panelGroupIdOrder.splice(idIndex, 1);
-
-      // Get usage of all remaining panel keys
-      const panelKeyMap = mapPanelToPanelGroups(draft.panelGroups);
-
-      // For the panel keys of the items that were just deleted, see if they're still used and if not, also delete the
-      // panel definition
-      for (const panelKey of panelKeys) {
-        const panelKeyUsage = panelKeyMap[panelKey];
-        if (panelKeyUsage === undefined || panelKeyUsage.length === 0) {
-          delete draft.panels[panelKey];
-        }
-      }
-    });
-  },
-
-  openDeletePanelGroupDialog: (panelGroupId) => {
-    const panelGroup = get().panelGroups[panelGroupId];
-    if (panelGroup === undefined) {
-      throw new Error(`Panel group with Id ${panelGroupId} not found`);
-    }
-    set((state) => {
-      state.deletePanelGroupDialog = { panelGroupId, panelGroupName: panelGroup.title };
-    });
-  },
-
-  closeDeletePanelGroupDialog: () =>
-    set((state) => {
-      state.deletePanelGroupDialog = undefined;
-    }),
 });
