+++ ui/dashboards/src/context/DashboardProvider/panel-group-slice.ts
