+++ ui/dashboards/src/context/DashboardProvider/panel-slice.ts
